+++ src/test/resources/inspections/CollectFunctionInCollectionInspection/nonMatches/collect_on_mixed_variable.php
@@ -1,0 +1,6 @@
+<?php
+
+/** @var mixed|array $var */
+$var = [];
+
+collect($var);